powf int_float_wide: (float)value drops frsp double path (80.39->84.35)

+++ src/dolphin/MSL_C/PPCEABI/bare/H/exponentialsf.c
@@ -52,7 +52,7 @@ static inline float int_float(s32 value)
 
 static inline float int_float_wide(s32 value)
 {
-    return (float)(double)value;
+    return (float)value;
 }
 
 static inline int classify_float(float value)
